DTO e Data Annotation

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,14 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using ProEventos.Persistence;
-using ProEventos.Domain;
-using ProEventos.Persistence.Contextos;
 using ProEventos.Application.Contratos;
 using Microsoft.AspNetCore.Http;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -30,7 +25,7 @@ namespace ProEventos.API.Controllers
           try
           {
               var eventos = await _eventoService.GetAllEventosAsync(true);
-              if (eventos == null) return NotFound("Nenhum evento encontrado.");
+              if (eventos == null) return NoContent();
 
               return Ok(eventos);
           }
@@ -48,7 +43,7 @@ namespace ProEventos.API.Controllers
             try
             { 
               var evento = await _eventoService.GetEventoByIdAsync(id, true);
-              if (evento == null) return NotFound("Evento por Id não encontrado.");
+              if (evento == null) return NoContent();
 
               return Ok(evento);
             }
@@ -66,7 +61,7 @@ namespace ProEventos.API.Controllers
             try
             { 
               var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-              if (evento == null) return NotFound("Evento por tema não encontrados.");
+              if (evento == null) return NoContent();
 
               return Ok(evento);
             }
@@ -79,12 +74,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             { 
                 var evento = await _eventoService.AddEventos(model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+                if (evento == null) return NoContent();
 
               return Ok(evento);
             }
@@ -96,12 +91,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             { 
                 var evento = await _eventoService.UpdateEvento(id, model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+                if (evento == null) return NoContent();
 
               return Ok(evento);
             }
@@ -117,9 +112,12 @@ namespace ProEventos.API.Controllers
         {
            try
             { 
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if (evento == null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ?
                        Ok("Deletado") :
-                       BadRequest("Evento não deletado");
+                       throw new Exception("Ocorreu um problema não específico ao tentar deletar o Evento");
             }
             catch (Exception ex)
             {
